move signed cookies into their own module
